move AnswerList component to so separate folder
